+++ app/_constants/form.ts
@@ -1,0 +1,1 @@
+export const GOOGLE_FORM_URL = 'https://forms.gle/HQBgGQDRiLj2r1jf8';
